Refactored spreading rate
