+++ covid19_inference/model/spreading_rate/change_points.py
@@ -1,279 +1,11 @@
-# ------------------------------------------------------------------------------ #
-# This file implements the change points in the spreading rate
-# ------------------------------------------------------------------------------ #
-
 import logging
 import pymc as pm
+import numpy as np
 import aesara.tensor as at
-
-import scipy.special
-from .model import *
-from . import utility as ut
+from .. import utility as ut
+from ..model import modelcontext
 
 log = logging.getLogger(__name__)
-
-
-def R_t_log_with_longtailed_dist(
-    R_0_log, dist=None, model=None, name_R_t=None, rho_scale_prior=0.1, **dist_priors
-):
-    r"""
-    Builds a time dependent reproduction number :math:`R_t` by using a long tailed distribution.
-
-    .. math::
-
-        log(R_t) &= log(R_0) + cumsum( s \cdot \Delta \rho) \\
-        s &\sim \text{HalfCauchy}(\mu=0.1)
-
-    The parameter :math:`\Delta \rho` is dependent on the given distribution. Possible values are
-    ``studentT``, ``weibull``, ``cauchy``. In theory you can also pass your another function of type
-    :class:`pm.distributions.Continuous`.
-
-
-    Parameters
-    ----------
-    R_0_log : number
-        Initial value of the reproduction number in log transform
-    dist : string or :class:`pm.distributions.Continuous`
-        Distribution to use. Possible values are ``studentT``, ``weibull``, ``cauchy``.
-        You can also pass your own distribution. If you use your own, be sure to configure the kwargs of the distribution
-        using the ``dist_priors`` parameter. The distribution should have the time dimension as first axis.
-    dist_priors : dict
-        Dictionary of kwargs i.e. mostly priors for the distribution.
-    model : :class:`Cov19Model`
-        if none, it is retrieved from the context
-
-    Returns
-    -------
-    R_t_log
-    """
-
-    log.info("R_t with long tailed dist")
-    # Get our default mode context
-    model = modelcontext(model)
-
-    # Parse dist priors
-    if "name" not in dist_priors:
-        dist_priors["name"] = "Delta_rho"
-    if "shape" not in dist_priors:
-        dist_priors["shape"] = model.sim_len
-
-    # Helper functions to define the distributions
-    def _get_StundentT_dist(**dist_priors):
-        if "nu" not in dist_priors:
-            dist_priors["nu"] = 1.5
-        if "mu" not in dist_priors:
-            dist_priors["mu"] = 0
-        if "sigma" not in dist_priors and "lam" not in dist_priors:
-            dist_priors["sigma"] = 1
-        return pm.StudentT(**dist_priors)
-
-    def _get_Weibull_dist(**dist_priors):
-        if "alpha" not in dist_priors:
-            dist_priors["alpha"] = 0.2
-        if "beta" not in dist_priors:
-            dist_priors["beta"] = 1
-        if "scale" not in dist_priors:
-            mean_influx = 1
-            scale = mean_influx / scipy.special.gamma(1 + 1 / dist_priors["alpha"])
-        else:
-            scale = dist_priors["scale"]
-            del dist_priors["scale"]
-        dist = (
-            (pm.Weibull(**dist_priors))
-            * scale
-            * (pm.Normal("wb_scale", 0, 1, shape=dist_priors["shape"]))
-        )
-        return dist
-
-    def _get_Cauchy_dist(**dist_priors):
-        if "alpha" not in dist_priors:
-            dist_priors["alpha"] = 0
-        if "beta" not in dist_priors:
-            dist_priors["beta"] = 1
-        return pm.Cauchy(**dist_priors)
-
-    # Parse dist
-    if dist is None:
-        dist = "studentT"
-    if isinstance(dist, str):
-        if dist == "studentT":
-            dist = _get_StundentT_dist(**dist_priors)
-        elif dist == "weibull":
-            dist = _get_Weibull_dist(**dist_priors)
-        elif dist == "cauchy":
-            dist = _get_Cauchy_dist(**dist_priors)
-    elif isinstance(dist, pm.distributions.Continuous):
-        dist = dist("", **dist_priors)
-    elif not isinstance(dist, pm.distributions.Continuous):
-        raise ValueError(f"Distribution {dist} is not supported")
-
-    # Build the model equations
-    scale = pm.HalfCauchy("Delta_rho_scale", beta=rho_scale_prior)
-    R_t_log = R_0_log + at.cumsum(scale * dist, axis=0)
-
-    # Create responding R_t pymc3 variable with given name (from parameters)
-    if name_R_t is not None:
-        pm.Deterministic(name_R_t, at.exp(R_t_log))
-
-    return R_t_log
-
-
-def lambda_t_with_sigmoids(
-    change_points_list,
-    pr_median_lambda_0,
-    pr_sigma_lambda_0=0.5,
-    model=None,
-    name_lambda_t="lambda_t",
-    hierarchical=None,
-    sigma_lambda_cp=None,
-    sigma_lambda_week_cp=None,
-    prefix_lambdas="",
-    shape=None,
-):
-    r"""
-    Builds a time dependent spreading rate :math:`\lambda_t` with change points. The change points are marked by a transient with a sigmoidal shape.
-
-    TODO
-    ----
-    Add a bit more detailed documentation.
-
-    Parameters
-    ----------
-    change_points_list
-    pr_median_lambda_0
-    pr_sigma_lambda_0
-    model : :class:`Cov19Model`
-        if none, it is retrieved from the context
-
-    Returns
-    -------
-    lambda_t_log
-    """
-    log.info("Lambda_t with sigmoids")
-    # Get our default mode context
-    model = modelcontext(model)
-    hierarchical = (
-        model.is_hierarchical if model.is_hierarchical is None else hierarchical
-    )
-    # ?Get change points random variable?
-    if shape is None:
-        shape = model.shape_of_regions
-    if isinstance(shape, int):
-        shape = (shape,)
-
-    lambda_log_list, tr_time_list, tr_len_list = _make_change_point_RVs(
-        change_points_list,
-        pr_median_lambda_0,
-        pr_sigma_lambda_0,
-        model=model,
-        hierarchical=hierarchical,
-        sigma_lambda_cp=sigma_lambda_cp,
-        sigma_lambda_week_cp=sigma_lambda_week_cp,
-        prefix_lambdas=prefix_lambdas,
-        shape=shape,
-    )
-
-    # Build the time-dependent spreading rate
-    lambda_log_t_list = [lambda_log_list[0] * at.ones((model.sim_len,) + shape)]
-    lambda_before = lambda_log_list[0]
-
-    # Loop over all lambda values and there corresponding transient values
-    for tr_time, tr_len, lambda_after in zip(
-        tr_time_list, tr_len_list, lambda_log_list[1:]
-    ):
-        # Create the right shape for the time array
-        t = np.arange(model.sim_shape[0])
-
-        # If the model is hierarchical repeatly add the t array to itself to match the shape
-        if shape:
-            t = np.repeat(t[:, None], shape, axis=-1)
-
-        # Applies standart sigmoid nonlinearity
-        lambda_t = at.sigmoid((t - tr_time) / tr_len * 4) * (
-            lambda_after - lambda_before
-        )  # tr_len*4 because the derivative of the sigmoid at zero is 1/4, we want to set it to 1/tr_len
-
-        lambda_before = lambda_after
-        lambda_log_t_list.append(lambda_t)
-
-    # Sum up all lambda values from the list
-    for i, lambda_t in enumerate(lambda_log_t_list):
-        pm.Deterministic(f"{prefix_lambdas}lambda_t_part_{i}", lambda_t)
-
-    lambda_t_log = sum(lambda_log_t_list)
-
-    # Create responding lambda_t pymc3 variable with given name (from parameters)
-    if name_lambda_t is not None:
-        pm.Deterministic(name_lambda_t, at.exp(lambda_t_log))
-
-    return lambda_t_log
-
-
-def lambda_t_with_linear_interp(
-    change_points_list,
-    pr_median_lambda_0,
-    pr_sigma_lambda_0=0.5,
-    model=None,
-    name_lambda_t="lambda_t",
-):
-    """
-    Parameters
-    ----------
-    change_points_list
-    pr_median_lambda_0
-    pr_sigma_lambda_0
-    model : :class:`Cov19Model`
-        if none, it is retrieved from the context
-
-    Returns
-    -------
-    lambda_t_log
-
-    TODO
-    ----
-    Documentation on this
-    """
-    log.info("Lambda_t linear in lin-space")
-    # Get our default mode context
-    model = modelcontext(model)
-
-    # ?Get change points random variable?
-    lambda_log_list, tr_time_list, tr_len_list = _make_change_point_RVs(
-        change_points_list, pr_median_lambda_0, pr_sigma_lambda_0, model=model
-    )
-
-    # Build the time-dependent spreading rate
-    lambda_t_list = [
-        at.exp(lambda_log_list[0]) * at.ones(model.sim_shape)
-    ]  # model.sim_shape = (time, state)
-    lambda_log_before = lambda_log_list[0]
-
-    # Loop over all lambda values and there corresponding transient values
-    for tr_time, tr_len, lambda_log_after in zip(
-        tr_time_list, tr_len_list, lambda_log_list[1:]
-    ):
-        # Create the right shape for the time array
-        t = np.arange(model.sim_shape[0])
-
-        # If the model is hierarchical repeatly add the t array to itself to match the shape
-        if model.is_hierarchical:
-            t = np.repeat(t[:, None], model.sim_shape[1], axis=-1)
-
-        step_t = at.clip((t - tr_time + tr_len / 2) / tr_len, 0, 1)
-
-        # create the time series of lambda
-        lambda_t = step_t * (at.exp(lambda_log_after) - at.exp(lambda_log_before))
-
-        lambda_log_before = lambda_log_after
-        lambda_t_list.append(lambda_t)
-
-    # Sum up all lambda values from the list
-    lambda_t_log = at.log(sum(lambda_t_list))
-
-    # Create responding lambda_t pymc3 variable with given name (from parameters)
-    pm.Deterministic(name_lambda_t, at.exp(lambda_t_log))
-
-    return lambda_t_log
 
 
 def _make_change_point_RVs(
@@ -562,7 +294,7 @@ def _make_change_point_RVs(
     )
 
     for cp_priors in change_points_list:
-        set_missing_priors_with_default(cp_priors, default_priors_change_points)
+        ut.set_missing_priors_with_default(cp_priors, default_priors_change_points)
 
     model = modelcontext(model)
     lambda_log_list = []
@@ -577,38 +309,3 @@ def _make_change_point_RVs(
         non_hierarchical_mod()
 
     return lambda_log_list, tr_time_list, tr_len_list
-
-
-def _smooth_step_function(start_val, end_val, t_begin, t_end, t_total):
-    """
-    Instead of going from start_val to end_val in one step, make the change a
-    gradual linear slope.
-
-    Parameters
-    ----------
-        start_val : float
-            Starting value
-
-        end_val : float
-            Target value
-
-        t_begin : int or array_like or :class:`~aesara.tensor.Variable`
-            Time point (inbetween 0 and t_total) where start_val is placed
-
-        t_end : int or array_like or :class:`~aesara.tensor.Variable`
-            Time point (inbetween 0 and t_total) where end_val is placed
-
-        t_total : int
-            Total number of time points
-
-    Returns
-    -------
-        : :class:`~aesara.tensor.Variable`
-            vector of length t_total with the values of the parameterised f(t)
-    """
-    t = np.arange(t_total)
-
-    return (
-        at.clip((t - t_begin) / (t_end - t_begin), 0, 1) * (end_val - start_val)
-        + start_val
-    )
